moving things to xy visualization
