+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/editor_frame.tsx
@@ -139,7 +139,6 @@ export function EditorFrame(props: EditorFrameProps) {
                 framePublicAPI={framePublicAPI}
                 uiActions={props.plugins.uiActions}
                 indexPatternService={props.indexPatternService}
-                eventAnnotationService={props.plugins.eventAnnotation}
               />
             </ErrorBoundary>
           )
