+++ x-pack/plugins/lens/public/mocks/store_mocks.tsx
@@ -47,7 +47,6 @@ export const defaultState = {
   query: { language: 'lucene', query: '' },
   resolvedDateRange: { fromDate: '2021-01-10T04:00:00.000Z', toDate: '2021-01-10T08:00:00.000Z' },
   isFullscreenDatasource: false,
-  isLoadLibraryVisible: false,
   isSaveable: false,
   isLoading: false,
   isLinkedToOriginatingApp: false,
