+++ x-pack/plugins/lens/public/state_management/types.ts
@@ -46,7 +46,6 @@ export interface EditorFrameState extends PreviewState {
   applyChangesCounter?: number;
   changesApplied?: boolean;
   isFullscreenDatasource?: boolean;
-  isLoadLibraryVisible?: boolean;
 }
 export interface LensAppState extends EditorFrameState {
   persistedDoc?: Document;
@@ -64,7 +63,6 @@ export interface LensAppState extends EditorFrameState {
   sharingSavedObjectProps?: Omit<SharingSavedObjectProps, 'sourceId'>;
   // Dataview/Indexpattern management has moved in here from datasource
   dataViews: DataViewsState;
-  isLoadLibraryVisible?: boolean;
 }
 
 export type DispatchSetState = (state: Partial<LensAppState>) => {
